tests: move `test_api` to `api.test_mpsk` and introduce class
